requirements.txt
boto3==1.29.6
python-dotenv==1.0.0
scikit-learn==1.3.0
pandas==1.3.4
requests==2.31.0